Create machinelearning.py
Implements Elastic Net machine learning model to predict the number of gas stations seen in the secretary phase of the trip.

Current data will look at 0.6-0.85 as the secretary phase with beta=0.75.

The CSV is structured: n(0-0.6), highwayLength, EstDev, n(0.6-0.85)

where n(...) is the number of gas stations on ... segment of highway.